feat(asset): cook localization tables and bridge them to the runtime catalog
Tina::Localization shipped with its contract complete and nothing connected:
no recipe verb, so no producer could cook AssetKind::LocalizationTable, and no
adapter, so the runtime could not reach a cooked table at all. Both halves now
exist.

The strings live in an external key=value file rather than inline in the recipe.
splitWs() cuts on whitespace and localized text almost always contains spaces;
teaching the tokenizer to quote would change every other verb's lexing, and
translator-authored content belongs in a file a translator can edit. The grammar
allows # comments, \n \t \ \= escapes, and unescaped = inside values (only the
first one separates, so URLs and format strings need no care). The key side is
trimmed and the value side is not: a trailing space can be deliberate.

Sorting by key hash is the cooker's job, not the author's. The wire requires
strictly ascending hashes, and hash order has no relation to alphabetical order,
so asking the author to pre-sort would be asking them to sort by a number they
cannot see. A duplicate key -- or a genuine 64-bit collision between two
distinct keys -- is rejected at cook time instead of letting one string silently
shadow the other at runtime.

The bridge lives on the Asset side and depends downward onto Tina::Localization,
mirroring TileMapNavigation2D -> Tina::Navigation2D. Localization links Core
only, so putting it there would invert the dependency. Both sides already share
Core::stringKeyHash; two copies of that hash would leave the table well formed
and make every lookup miss silently.

Records one wire/runtime inconsistency without papering over it: the runtime's
AbsentTextOffset sentinel can never arrive from a cooked payload, because the
wire validator rejects any textOffset past textBytes. A locale meaning to blank
a string authors an empty value, which resolves as an empty string.

Also unblocks two test targets that did not compile at HEAD. The mesh index
U16->U32 migration removed StaticMeshIndexType::U16 from the enum and changed
makeCanonicalUnitCubeMeshDesc to take span<u32>, but left eight asset and
asset_format fixtures behind -- 19 errors across 5 files, and
tina_asset_format_tests had never produced an exe. Nine index arrays and four
enum references are migrated. jointIndices/jointWeights stay u16 on purpose:
they are fixed-point skin influences, and SkinnedMeshPayloadTests.cpp:240 is a
12-element joint array confusingly named "indices", so a rename by name would
have silently corrupted skinning data.

Evidence: new LocalizationTableLoadTests 12/12 including a cook -> payload ->
catalog round trip that resolves the same strings by the same keys;
tina_localization_tests 20/20; tina_asset_format_tests 148/148, up from not
building; tina_asset_tests compiles and 382 pass.

Nine failures remain in tina_asset_tests, all in retirement/GPU-owner paths this
commit does not touch -- two of them in files it does not modify at all. They
were unobservable while the target could not compile and are tracked as
ASSET-TEST-U32-INDEX-DEBT-001; whether each is a source defect or an expectation
stale since the U32 migration still needs deciding, and expectations must not be
edited to force green.

+++ tests/asset/AssetGpuMeshTests.cpp
@@ -28,7 +28,7 @@ TEST(AssetGpuMeshTests, UploadTypedStaticMeshToNullDevice)
 
     std::array<AssetFormat::StaticMeshSubmeshDesc, 1> submeshes{};
     std::array<float, 24 * AssetFormat::StaticMeshWire::FloatsPerVertex> vertices{};
-    std::array<Core::u16, 36> indices{};
+    std::array<Core::u32, 36> indices{};
     const auto desc = AssetFormat::makeCanonicalUnitCubeMeshDesc(submeshes, vertices, indices);
     ASSERT_FALSE(desc.vertices.empty());
 
@@ -57,7 +57,7 @@ TEST(AssetGpuMeshTests, RejectsZeroMeshKey)
     const auto meshId = *Core::AssetId::fromBytes(idBytes(2U));
     std::array<AssetFormat::StaticMeshSubmeshDesc, 1> submeshes{};
     std::array<float, 24 * AssetFormat::StaticMeshWire::FloatsPerVertex> vertices{};
-    std::array<Core::u16, 36> indices{};
+    std::array<Core::u32, 36> indices{};
     const auto desc = AssetFormat::makeCanonicalUnitCubeMeshDesc(submeshes, vertices, indices);
     auto cooked = AssetFormat::writeCookedStaticMeshAsset(meshId, desc);
     ASSERT_TRUE(cooked.has_value());
@@ -82,7 +82,7 @@ TEST(AssetGpuMeshTests, UploadsStaticMeshToNullDevice)
         1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 1, 0,
         0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1,
     };
-    const std::array<Core::u16, 3> indices{0, 1, 2};
+    const std::array<Core::u32, 3> indices{0, 1, 2};
     auto cooked = AssetFormat::writeCookedStaticMeshAsset(meshId, AssetFormat::StaticMeshPayloadDesc{
         .boundsCenterX = 0.5F,
         .boundsCenterY = 0.5F,
